Remove unused BaseParameter
